requirements.txt
matplotlib==3.5.1
numpy==1.21.5
torchvision==0.12.0
pandas==1.4.1
pillow==9.0.1
requests==2.27.1
joblib==1.1.0